pyproject.toml
[project]
name = "subsidy"
version = "0.1.0"
description = "地方自治体助成金 自動収集・要約・公開システム"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "pdfplumber>=0.11.0",
    "pymupdf>=1.24.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "tenacity>=8.3.0",
    "pydantic>=2.7.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.1",
    "lxml>=5.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
